refactor(agent): extract the queue and resume coordinators from AgentRuntime
QueuedRunCoordinator takes enqueue() and runQueued(). The two halves are
one protocol and only make sense together: what enqueue() stores — a
serialised request on a fail-closed row, dispatched only after the row
exists — is what runQueued() has to claim, position and rehydrate. Both
fail closed in the same direction, so a queued run that cannot be stored,
dispatched, positioned or rehydrated is settled rather than left as an
orphan a worker would later find in an impossible state.

ResumeCoordinator takes approve() and submitInput(). Both follow the same
claim protocol, and the order in it is the safety property: probe that a
resume handle can be built, win the atomic claim, then re-resolve the
event-stream position from a fresh row because the claim moved it.
ADR-105's deliberate divergence — an input is validated against the
tool's schema before anything is claimed, so a rejection leaves the run
resumable — is now a divergence inside one class rather than a difference
between two long methods to be spotted by reading both.

Both were only separable once the executor became injectable: each ends
by driving the run, and neither needs a DTO or a widened return type to
hand that work on.

They are one commit because their changes to AgentRuntime interleave;
splitting them would mean committing an intermediate state that was never
run through the gates.

AgentRuntime is 266 lines, down from 1218. What remains is the interface
contract, the published constants, the retry predicate and delegation.
cancel(), events() and status() stay: their guard is already a method on
AiActorContext, so there is nothing left in them to extract.

The roadmap item is removed rather than marked shipped — the file says
outright that anything shipped belongs in the changelog.
