+++ ROADMAP.md
@@ -16,24 +16,18 @@ governance audit trail.
 
 ## Next — agent runtime maturity
 
-1. **Decompose the agent runtime into focused internal services.**
-   `AgentRuntime` has grown through the hardening arc; split it along its
-   natural seams (authorization, request codec, queue coordination, lease
-   management, resume coordination, run execution, failure recovery, outcome
-   mapping) while staying one package (ADR-090). Behaviour-preserving; makes
-   the runtime testable per concern and safe to extend.
-2. **Make the tool loop's security collaborators mandatory.** The
+1. **Make the tool loop's security collaborators mandatory.** The
    `ToolLoopService` gates (tool-call policy, per-configuration allow-list,
    input schema validation) are optional constructor arguments today — absent
    wiring silently weakens the gate. Make them required, and give tests an
    explicit lean wiring (`Null*` implementations plus a `Testing` builder)
    instead of implicit absence. A container test pins the production wiring.
-3. **Conversation-level context-window management.** The tool loop already
+2. **Conversation-level context-window management.** The tool loop already
    bounds its transcript (ADR-107); conversations do not. Plan against the
    smallest reachable model window in the fallback chain — or re-fit when a
    fallback actually fires — so a long conversation degrades predictably
    instead of failing on the provider.
-4. **A first-class contract for side-effecting tools.** Promote the tool
+3. **A first-class contract for side-effecting tools.** Promote the tool
    effect declaration (read-only / idempotent write / non-idempotent write,
    ADR-111) into a tool-facing interface with an idempotency scope and an
    optional preview, so writing tools can be built against a contract instead
