+++ Tests/Unit/Service/Agent/AgentRuntimeTest.php
@@ -44,7 +44,9 @@ use Netresearch\NrLlm\Service\Agent\Exception\RunNotAwaitingInputException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunStateUnavailableException;
 use Netresearch\NrLlm\Service\Agent\InputSubmission;
 use Netresearch\NrLlm\Service\Agent\Queue\AgentRunQueuedMessage;
+use Netresearch\NrLlm\Service\Agent\QueuedRunCoordinator;
 use Netresearch\NrLlm\Service\Agent\QueuedRunFailureRecovery;
+use Netresearch\NrLlm\Service\Agent\ResumeCoordinator;
 use Netresearch\NrLlm\Service\Option\ToolOptions;
 use Netresearch\NrLlm\Service\Tool\ActingBackendUserResolverInterface;
 use Netresearch\NrLlm\Service\Tool\AgentRunPersister;
@@ -70,6 +72,8 @@ use Throwable;
 #[CoversClass(AgentRuntime::class)]
 #[CoversClass(AgentRunExecutor::class)]
 #[CoversClass(QueuedRunFailureRecovery::class)]
+#[CoversClass(QueuedRunCoordinator::class)]
+#[CoversClass(ResumeCoordinator::class)]
 final class AgentRuntimeTest extends AbstractUnitTestCase
 {
     private RecordingAgentRunRepository $repository;
